remove unnecessary cast

+++ src/autopas/sph/SPHParticle.h
@@ -413,7 +413,7 @@ class SPHParticle : public autopas::Particle {
    * @return Value of the requested attribute.
    */
   template <AttributeNames attribute>
-  constexpr typename std::tuple_element<static_cast<size_t>(attribute), SoAArraysType>::type::value_type get() {
+  constexpr typename std::tuple_element<attribute, SoAArraysType>::type::value_type get() {
     if constexpr (attribute == AttributeNames::ptr) {
       return this;
     } else if constexpr (attribute == AttributeNames::mass) {
@@ -461,8 +461,7 @@ class SPHParticle : public autopas::Particle {
    * @param value New value of the requested attribute.
    */
   template <AttributeNames attribute>
-  constexpr void set(
-      typename std::tuple_element<static_cast<size_t>(attribute), SoAArraysType>::type::value_type value) {
+  constexpr void set(typename std::tuple_element<attribute, SoAArraysType>::type::value_type value) {
     if constexpr (attribute == AttributeNames::mass) {
       setMass(value);
     } else if constexpr (attribute == AttributeNames::posX) {
